+++ src/main/remote.c
@@ -36,9 +36,9 @@ struct remote_msg {
 	bool bytes;			 /* bytes transferred flag */
 	bool bufpercent;		 /* transfer buffer percentage flag */
 	unsigned int lastwritten;	 /* last-written bytes count */
-	size_t rate_limit;		 /* rate limit, in bytes per second */
+	off_t rate_limit;		 /* rate limit, in bytes per second */
 	size_t buffer_size;		 /* buffer size, in bytes (0=default) */
-	size_t size;			 /* total size of data */
+	off_t size;			 /* total size of data */
 	double interval;		 /* interval between updates */
 	unsigned int width;		 /* screen width */
 	unsigned int height;		 /* screen height */
